Fixed create job installer bug & mismatched query
Fixed a bug with how create job was handling the dropdown menu being set to n/a and allowed it to select a installer with no billed.

Fixed the prisma query for finding mismatched jobs.

Changed the layout of view jobs, moved buttons to top and dropdowns to middle to reflect intended functionality