Form clearing Bugfix and user input fix
Fixed a bug that prevented the store text box from clearing when the user successfully created a job.

Fixed a bug that let a user delete both billed and paid from the same job. Now jobs still require one or the other before allowing the user to save the edit.